Fix problem, where it was not parsing optional parameters correctly
If a route has an optional parameter, the regex building routine in the parse
phase of the router incorrectly failed to add a bracketed test to the regex it
is constructing to be able to parse all valid urls.  However it did increase the
counter.  So when we came to match a route that was after (in the parse list) of
valid routes, it wasn't finding it, because the match position was off.