Considerable change to router to make it work with a more complex dom structure
Changes in this release

Replacing forEach function calls with for(item of items) so it makes stepping
through the code much easier

Adding support for an event 'akc-route-changed' with event.detail being the new
url that was routed too.

Stopping dispatching until the dom parsing is ready.

A (untested) reset function which will reparse the dom (should it have changed).
It is because the dom may be being reparsed that we hold up dispatching whilst
it is.

Bug fixes to the handling of routers with children - using them as a way of
having a small area of a page respond to sub parts of a url. Let me explain...

First of all, the dom is now parsed properly going into custom elements and
parsing the effective dom "AS IT WILL BE DISTRIBUTED".  In other words, the
content pulled in by the <content></content> tag into the local DOM of the
element as it is then stamped into the DOM of the page is assumed to exist in
the correct place.

Secondly, a custom element may both be a target of a route (including AKC.Route
behavior) and have children.  This allows you to have a route like /reports
which at the top level is a target to be displayed via a <iron-pages> (or
similar construct) and then inside itself have a home route to display a
selection page, and other subject routes which display complete pages in their
own right.  The selections will work at both levels

Thirdly, at the internal level an <akc-route-selector> without children can be used to
select amongst its other routeable siblings.  But an <akc-route-selector> with
children will not try to select its siblings, but instead assumes a small part
of the elements real estate is reserved for swapping between the content of its
child routeable elements.

So for instance /appointments/20160124 might bring up an appointments page for
today. This is a custom element.  Inside it might have a home page listing all
the offices and other pages for each of the separate offices.  These would be
selected via /appointments/20160124/london or /appointments/20160124/brighton.
But also, there might be functionaity common to all pages to book and
appointment, or alter an existing appointment.  Assuming booking was default
(sub url /) then we might also want /appointment/2016124/edit, or
/appointment/20160124/london/edit.

By using an <akc-route-selector> for the office element at the other main page
swapping function for the majority functionality for either booking or editing
(or other activity) we can change both independantly within the same dom
structure but with different urls.

All the above was planned to be available but wasn't working.  Now it has tests
and it works.